+++ data/src/main/java/com/umc/data/dataSource/remote/attendance/AttendanceRemoteDataSource.kt
@@ -8,7 +8,7 @@ import com.umc.domain.model.request.attendance.AttendanceCheckRequest
 interface AttendanceRemoteDataSource {
     suspend fun getAttendanceAvailable(): ApiState<List<AttendanceAvailableResponse>>
     suspend fun postAttendanceCheck(request: AttendanceCheckRequest): ApiState<String>
-    suspend fun getPendingUsers(scheduleId: Int): ApiState<List<AdminPendingUserResponse>>
-    suspend fun approveAttendance(recordId: Int): ApiState<Unit>
-    suspend fun rejectAttendance(recordId: Int): ApiState<Unit>
+    suspend fun getPendingUsers(scheduleId: Long): ApiState<List<AdminPendingUserResponse>>
+    suspend fun approveAttendance(recordId: Long): ApiState<Unit>
+    suspend fun rejectAttendance(recordId: Long): ApiState<Unit>
 }
